Start Jersey and Swagger Unit Tests
