+++ restfriends-swagger-jersey1/src/test/java/com/github/bordertech/swagger/test/Jersey1ApiTest.java
@@ -82,17 +82,4 @@ public class Jersey1ApiTest {
 				.statusCode(500);
 	}
 
-	@Test
-	public void testLaunchSwagger() {
-		RestAssured.given().
-				log().all()
-				.when()
-				.get("/launchswagger")
-				.then()
-				.assertThat()
-				.statusCode(200)
-				.and()
-				.contentType(ContentType.HTML)
-				.body("html.head.title", IsEqual.equalTo("Swagger UI"));
-	}
 }
